# FlightBookingApp/Backend/src/docs/swagger.yaml
definitions:
...
  dto.CreatedResponse:
    properties:
      id:
        type: string
    type: object
...
  dto.SimpleResponse:
    properties:
      message:
        type: string
    type: object
  dto.UserInfo:
    properties:
      address:
        $ref: '#/definitions/model.Address'
      name:
        type: string
      surname:
        type: string
    required:
    - address
    - name
    - surname
    type: object
...
  model.Address:
    properties:
      city:
        example: Novi Sad
        type: string
      country:
        example: Serbia
        type: string
      street:
        example: Rumenacka
        type: string
      streetNumber:
        example: 21a
        type: string
    required:
    - city
    - country
    - street
    - streetNumber
    type: object
  model.Airport:
    properties:
      address:
        $ref: '#/definitions/model.Address'
      id:
        example: 641d71cedd5e51a214a40c14
        type: string
      name:
        example: Nikola Tesla
        type: string
    required:
    - address
    - name
    type: object
...
  model.User:
    properties:
      address:
        $ref: '#/definitions/model.Address'
      id:
        type: string
      name:
        type: string
      surname:
        type: string
    required:
    - address
    - name
    - surname
    type: object
...
info:
  contact: {}
paths:
...
  /airport:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Airport'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      tags:
      - Airport
  /airport/{id}:
    get:
      parameters:
      - description: Airport ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Airport'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      tags:
      - Airport
...
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      security:
      - bearerAuth: []
      tags:
      - User
    post:
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      security:
      - bearerAuth: []
      tags:
      - User
  /user/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      security:
      - bearerAuth: []
      tags:
      - User
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.SimpleResponse'
      security:
      - bearerAuth: []
      tags:
      - User
  /user/logged-in:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "500":
          description: can't find your account info
          schema:
            type: string
      security:
      - bearerAuth: []
      tags:
      - User
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"